agents/requirements.txt
pyserial==3.5
pyzmq==27.1.0
flatbuffers==24.3.25
